refactor: writing the /portal on routes that are in this folder
