+++ frontend/src/functions/useSignIn.ts
@@ -19,7 +19,7 @@ export default function useSignIn() {
       if (response.status === 200) {
         const userToken = response.data.token as string
         setToken(userToken)
-        router.push('/dashboard')
+        router.push('/portal/dashboard')
       }
     } catch (error: unknown) {
       if (axios.isAxiosError(error)) {
